requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
supabase>=2.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
httpx>=0.24.0,<0.25.0
python-dateutil>=2.8.0
